smart-candidate-finder/requirements.txt
openai
python-dotenv
google-search-results
beautifulsoup4
requests